Refactor: Move AccessManager to separate file
